requirements.txt
requests
pandas
sqlalchemy
numpy
python-dotenv
pyodbc